@@ -13,7 +13,7 @@
     "utils"
   ],
   "scripts": {
-    "rollup": "del-cli -f dist && rollup -c --cjs2015 --esm2015 --dts",
+    "rollup": "del-cli -f dist && rollup -c",
     "rollup:watch": "npm run rollup -- --watch",
     "sync-pkg": "node build/rollup/_syncPkg.js --run-sync-pkg",
     "doctoc": "doctoc --github README.md",
@@ -35,50 +35,49 @@
     "@alorel/eslint-config-typescript": "^1.0.17",
     "@alorel/rollup-plugin-copy": "^1.0.2",
     "@alorel/rollup-plugin-copy-pkg-json": "^1.0.3",
-    "@alorel/rollup-plugin-threaded-terser": "^1.1.0",
-    "@rollup/plugin-commonjs": "^22.0.2",
-    "@rollup/plugin-node-resolve": "^14.0.0",
-    "@rollup/plugin-replace": "^4.0.0",
+    "@rollup/plugin-commonjs": "^23.0.2",
+    "@rollup/plugin-node-resolve": "^14.1.0",
     "@semantic-release/changelog": "^6.0.1",
     "@semantic-release/exec": "^6.0.3",
     "@semantic-release/git": "^10.0.1",
     "@semantic-release/github": "^8.0.6",
     "@semantic-release/npm": "^9.0.1",
     "@types/chai": "^4.3.3",
-    "@types/lodash": "^4.14.184",
-    "@types/mocha": "^9.1.1",
-    "@types/node": "^18.7.15",
+    "@types/lodash": "^4.14.188",
+    "@types/lodash-es": "^4.17.6",
+    "@types/mocha": "^10.0.0",
+    "@types/node": "^18.11.9",
     "@types/pluralize": "^0.0.29",
     "@types/sinon": "^10.0.13",
     "@types/uuid": "^8.3.4",
     "@types/zen-observable": "^0.8.3",
-    "@typescript-eslint/eslint-plugin": "^5.36.2",
-    "@typescript-eslint/parser": "^5.36.2",
+    "@typescript-eslint/eslint-plugin": "^5.42.0",
+    "@typescript-eslint/parser": "^5.42.0",
     "chai": "^4.3.6",
     "cross-env": "^7.0.3",
     "del-cli": "^5.0.0",
-    "doctoc": "^2.2.0",
-    "eslint": "^8.23.0",
-    "lazy-get-decorator": "^2.2.0",
-    "mocha": "^10.0.0",
+    "doctoc": "^2.2.1",
+    "eslint": "^8.26.0",
+    "lazy-get-decorator": "^2.2.1",
+    "mocha": "^10.1.0",
     "nyc": "^15.1.0",
     "pluralize": "^8.0.0",
-    "rollup": "^2.79.0",
-    "rollup-plugin-typescript2": "^0.33.0",
-    "rxjs": "^7.5.6",
+    "rollup": "^2.79.1",
+    "rollup-plugin-typescript2": "^0.34.1",
+    "rxjs": "^7.5.7",
     "semantic-release": "^19.0.5",
-    "sinon": "^14.0.0",
+    "sinon": "^14.0.1",
     "source-map-support": "^0.5.21",
-    "terser": "^5.15.0",
     "ts-node": "^10.9.1",
-    "tslib": "^2.4.0",
-    "typedoc": "~0.15.3",
-    "typescript": "^4.8.2",
+    "tslib": "^2.4.1",
+    "typedoc": "~0.15.8",
+    "typescript": "^4.8.4",
     "uuid": "^9.0.0",
-    "zen-observable": "^0.8.15"
+    "zen-observable": "^0.9.0"
   },
   "dependencies": {
-    "lodash": "^4.0.0"
+    "lodash": "^4.17.21",
+    "lodash-es": "^4.17.21"
   },
   "peerDependencies": {
     "rxjs": "^7.0.0"
